seller config debugation 80% without schedule

+++ Snappfood/database/migrations/2022_06_29_101728_create_rest_addresses_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('rest_addresses', function (Blueprint $table) {
             $table->id();
             $table->string("address");
-            $table->integer("latitude");
-            $table->integer("longitude");
+            $table->double('latitude', 8, 6);
+            $table->double('longitude', 8, 2);
             $table->timestamps();
         });
     }
